Switch from reflective Java to functional Java
okay so first i realized that i never used done_testing()
to check the BIG sum (instead this only checks little sums
and (through diag() calls) indirectly *suggest* that the
test *might* have skipped/overdone some test cases).  so
i had to remember to do that.... and then i looked at this
plan thing and thought... eh, we could take it out.

but then suddenly i'm left with this constructor fail that
could potentially skip the entire test so i had to do with
the possibility of that failing.... but then i realized
"wait a second, why should i even have to pass a class when
i'm just calling the constructor to make instances of it?"

then i thought of passing the constructor directly... and
the thought of passing a constructor led to my ultimate
realization that i could have just passed an arbitrary
function for this and that method/constructor literals
(whatever they're actually called) like ::new exists o_o

it's amazing how Java 1.8 is surprisingly "modern"....

* * *

oh yeah i also updated some stuff here and there, failing() is
removed for good (it does too many things and i don't like it
(never liked it much anyways....)) there is also a remark deleted
that is far from accurate now that we have moved away from a mono-
lithic exec*() call; there is, in a sense, more freedom with what
we can do with the exit status pertaining to the decision between
running this as a part of prove (that is to not set an exit status
unless we ourselves fail) vs running as a standalone program
(which on the contrary NEEDS to set error status accurately)
all in all this is taking too much time aeughhh
but it's pretty good i guess!

(runff is not my favorite either for the record. it might
just be my inability to get jdb(1) working, but it is
kind of useless....)